# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GAuth Open Core API — Mandate lifecycle, governance profiles, and credential management.
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MgmtHealthResponse {
  status: string;
  version: string;
  sdk: string;
  uptime_seconds: number;
}

export type MandateStatus = (typeof MandateStatus)[keyof typeof MandateStatus];

export const MandateStatus = {
  DRAFT: "DRAFT",
  ACTIVE: "ACTIVE",
  SUSPENDED: "SUSPENDED",
  REVOKED: "REVOKED",
  EXPIRED: "EXPIRED",
} as const;

export type GovernanceProfile =
  (typeof GovernanceProfile)[keyof typeof GovernanceProfile];

export const GovernanceProfile = {
  minimal: "minimal",
  standard: "standard",
  strict: "strict",
  enterprise: "enterprise",
  behoerde: "behoerde",
} as const;

export interface ToolPolicy {
  allowed?: boolean;
  requires_approval?: boolean;
  max_per_session?: number | null;
}

export interface PlatformPermissions {
  shell?: boolean;
  db_read?: boolean;
  db_write?: boolean;
  secrets_read?: boolean;
}

export interface MandateParties {
  grantor: string;
  subject: string;
  intermediary?: string;
}

export type MandateScopeCoreVerbs = { [key: string]: ToolPolicy };

export interface MandateScope {
  core_verbs?: MandateScopeCoreVerbs;
  platform_permissions?: PlatformPermissions;
  allowed_sectors?: string[];
  allowed_regions?: string[];
  allowed_decisions?: string[];
  allowed_paths?: string[];
  denied_paths?: string[];
}

export interface MandateBudget {
  total_cents?: number | null;
  remaining_cents?: number | null;
}

export interface MandateDelegation {
  parent_mandate_id?: string | null;
  delegation_depth?: number;
  delegate_agent_id?: string;
  max_depth?: number;
}

export interface MandateResponse {
  mandate_id: string;
  status: MandateStatus;
  governance_profile: GovernanceProfile;
  parties: MandateParties;
  scope: MandateScope;
  budget?: MandateBudget;
  delegation?: MandateDelegation;
  valid_from?: string | null;
  valid_until?: string | null;
  created_at: string;
  updated_at: string;
}

export interface CreateMandateRequest {
  governance_profile?: GovernanceProfile;
  parties: MandateParties;
  scope: MandateScope;
  budget?: MandateBudget;
  delegation?: MandateDelegation;
  valid_from?: string;
  valid_until?: string;
}

export interface ReasonRequest {
  reason: string;
}

export interface MandateListResponse {
  items: MandateResponse[];
  next_cursor?: string | null;
  total: number;
}

export type AuditEntryMetadata = { [key: string]: unknown };

export interface AuditEntry {
  id: string;
  mandate_id: string;
  action: string;
  actor?: string | null;
  reason?: string | null;
  previous_status?: string | null;
  new_status?: string | null;
  metadata?: AuditEntryMetadata;
  timestamp: string;
}

export interface AuditHistoryResponse {
  entries: AuditEntry[];
}

export interface DelegationChainEntry {
  mandate_id: string;
  delegation_depth: number;
  delegate_agent_id: string;
  parent_mandate_id?: string | null;
  budget_reserved_cents: number;
}

export interface DelegationChainResponse {
  chain: DelegationChainEntry[];
}

export interface ProfileCeilings {
  min_approval_mode: string;
  agent_delegation: boolean;
  max_delegation_depth: number;
  max_session_duration_minutes: number;
  max_tool_calls: number;
  shell_mode: string;
  db_production: boolean;
  db_write: boolean;
  secrets_read: boolean;
}

export type VciIssuerMetadataCredentialConfigurationsSupported = {
  [key: string]: { [key: string]: unknown };
};

export interface VciIssuerMetadata {
  credential_issuer: string;
  credential_endpoint: string;
  token_endpoint: string;
  credential_configurations_supported: VciIssuerMetadataCredentialConfigurationsSupported;
}

export type ListMandatesParams = {
  limit?: number;
  cursor?: string;
  status?: MandateStatus;
  governance_profile?: GovernanceProfile;
};