# lib/api-zod/src/generated/types/auditEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GAuth Open Core API — Mandate lifecycle, governance profiles, and credential management.
 * OpenAPI spec version: 0.1.0
 */
import type { AuditEntryMetadata } from "./auditEntryMetadata";

export interface AuditEntry {
  id: string;
  mandate_id: string;
  action: string;
  actor?: string | null;
  reason?: string | null;
  previous_status?: string | null;
  new_status?: string | null;
  metadata?: AuditEntryMetadata;
  timestamp: Date;
}