# lib/api-zod/src/generated/types/auditHistoryResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GAuth Open Core API — Mandate lifecycle, governance profiles, and credential management.
 * OpenAPI spec version: 0.1.0
 */
import type { AuditEntry } from "./auditEntry";

export interface AuditHistoryResponse {
  entries: AuditEntry[];
}