# lib/api-zod/src/generated/types/createMandateRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GAuth Open Core API — Mandate lifecycle, governance profiles, and credential management.
 * OpenAPI spec version: 0.1.0
 */
import type { GovernanceProfile } from "./governanceProfile";
import type { MandateBudget } from "./mandateBudget";
import type { MandateDelegation } from "./mandateDelegation";
import type { MandateParties } from "./mandateParties";
import type { MandateScope } from "./mandateScope";

export interface CreateMandateRequest {
  governance_profile?: GovernanceProfile;
  parties: MandateParties;
  scope: MandateScope;
  budget?: MandateBudget;
  delegation?: MandateDelegation;
  valid_from?: Date;
  valid_until?: Date;
}