# lib/api-zod/src/generated/types/governanceProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GAuth Open Core API — Mandate lifecycle, governance profiles, and credential management.
 * OpenAPI spec version: 0.1.0
 */

export type GovernanceProfile =
  (typeof GovernanceProfile)[keyof typeof GovernanceProfile];

export const GovernanceProfile = {
  minimal: "minimal",
  standard: "standard",
  strict: "strict",
  enterprise: "enterprise",
  behoerde: "behoerde",
} as const;