# lib/api-zod/src/generated/types/mandateDelegation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GAuth Open Core API — Mandate lifecycle, governance profiles, and credential management.
 * OpenAPI spec version: 0.1.0
 */

export interface MandateDelegation {
  parent_mandate_id?: string | null;
  delegation_depth?: number;
  delegate_agent_id?: string;
  max_depth?: number;
}