# lib/api-zod/src/generated/types/mandateStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GAuth Open Core API — Mandate lifecycle, governance profiles, and credential management.
 * OpenAPI spec version: 0.1.0
 */

export type MandateStatus = (typeof MandateStatus)[keyof typeof MandateStatus];

export const MandateStatus = {
  DRAFT: "DRAFT",
  ACTIVE: "ACTIVE",
  SUSPENDED: "SUSPENDED",
  REVOKED: "REVOKED",
  EXPIRED: "EXPIRED",
} as const;