# lib/api-zod/src/generated/types/mgmtHealthResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GAuth Open Core API — Mandate lifecycle, governance profiles, and credential management.
 * OpenAPI spec version: 0.1.0
 */

export interface MgmtHealthResponse {
  status: string;
  version: string;
  sdk: string;
  uptime_seconds: number;
}