# lib/api-zod/src/generated/types/profileCeilings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GAuth Open Core API — Mandate lifecycle, governance profiles, and credential management.
 * OpenAPI spec version: 0.1.0
 */

export interface ProfileCeilings {
  min_approval_mode: string;
  agent_delegation: boolean;
  max_delegation_depth: number;
  max_session_duration_minutes: number;
  max_tool_calls: number;
  shell_mode: string;
  db_production: boolean;
  db_write: boolean;
  secrets_read: boolean;
}