# lib/api-zod/src/generated/types/vciIssuerMetadata.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GAuth Open Core API — Mandate lifecycle, governance profiles, and credential management.
 * OpenAPI spec version: 0.1.0
 */
import type { VciIssuerMetadataCredentialConfigurationsSupported } from "./vciIssuerMetadataCredentialConfigurationsSupported";

export interface VciIssuerMetadata {
  credential_issuer: string;
  credential_endpoint: string;
  token_endpoint: string;
  credential_configurations_supported: VciIssuerMetadataCredentialConfigurationsSupported;
}